scraper/requirements.txt
requests
pymongo==3.10.1
python-dotenv==0.12.0
schedule
